# Attempt to improve SDRAM timing