# Overview 

Currently SDRAM timing doen't seem to be constrained at all.
All clocks are explicitly decoupled with each other, despite `sys_clk <-> sdram_clk` (and `sys_clk <-> cpu_clk`) should not be.

https://github.com/MiSTer-devel/Arcade-Cave_MiSTer/blob/26f0888ec7cd70f8b18c6ab912c918a6d3789f40/quartus/rtl/sdram.sdc#L6-L11

Thus, it seems that the following important constraints are ignored by Fitter and Timing Analyzer due to this decoupling.

https://github.com/MiSTer-devel/Arcade-Cave_MiSTer/blob/26f0888ec7cd70f8b18c6ab912c918a6d3789f40/quartus/rtl/sdram.sdc#L13-L30

I fixed constraints and explored a better phase shift.
I'm not sure but it might fix #44.

In a later section, I will show the results of some experiments and tests I did.
Note that I tested only with my 128 MB board and my Cyclone V chip.
I cannot perfectly rule out the possibility of timing issues on other 128 MB, 64 MB, 32 MB boards, other Cyclone V chips.

# Note and suggestion

Even with my changes, Timing Analyzer reports negative slack a bit on some of outputting `SDRAM_A` lines(Slow corner 85C, 0C).
I tried to minimize it but this can't be eliminated with the current design.

The problem is that there are combinational logics between registers and `SDRAM_A`(and also `SDRAM_BA` and `SDRAM_CKE`) pins.
Thus these registers cannot be assigned to fast output registers on IO block.
This is increasing clock-to-output time and is the main cause of the negative slack.

I added some location assignments to minimize delay, even without fast output registers.
It worked for some of `SDRAM_A` lines, and `SDRAM_BA` , `SDRAM_CKE` but it is not perfect but just temporary tweaks.

`SDRAM_A` lines share common source registers(`stateReg.011`, `stateReg.100`, `stateReg.101`) which cannot be duplicated well by Fitter probably due to logic complexity.
And the location of `SDRAM_A` pins are not bundled, but are spread throughout the chip, making routing worse.
Ideally, SDRAM controller should be fixed to have registers on output of `SDRAM_A`, `SDRAM_BA`, `SDRAM_CKE` so that they can be assigned to fast output registers on IO block.

# About negative slack

Of course I'm not sure if we can completely trust the provided slack value in case of interfacing with another device outside of the chip through IO.
Propagation delay is defined in constraints but it's just an assumption.
Clock delay outside the chip doesn't seem to be considered(and I don't know how to define it. Using virtual clock?).
But I assume we can use it as a reference to some extent.

# Results of experiments

All experiments were done with DDP.

## Prepareation

I cannot reproduce the issue caused by delay, since unfortunately(or fortunately?) my 128 MB SDRAM board seems to be relatively fast and/or my Cyclone V chip also seems to be relatively fast.
So I put 22 pF capacitor between `SDRAM_DQ[1]` and `GND`. 

Before putting it, my 128 MB SDRAM run up to 145 MHz on memtest core.
After putting it, the max freq reduced to 110 MHz.

Even after putting it, no fails detected at 110 MHz, 100 MHz, 90 MHz.
And no fails on jotego's sdram test core at 96 MHz.

Thus, Cave core should also work at 96 MHz with the capacitor mod.

I used this modded SDRAM board as "Slow Case".
Unmodded SDRAM board is referred to as "Normal Case".


## Experiment 1: Test on current source and constraints

### Procedure
- No modifiction to current source and constraints (used 26f0888)
- Made 20 builds with diffrent fitter seed value using DSE II.

### Result

| No  | Normal Case | Slow Case | Failure detail
| --- | ----------- | --------- | --------------
| 1   | OK          | Fail      | Crash and freeze at high score screen
| 2   | OK          | Fail      | Graphical glitch and crash, reset at title screen
| 3   | OK          | Fail      | Graphical glitch and crash, reset at title screen
| 4   | OK          | Fail      | Freeze after ATLUS logo
| 5   | OK          | Fail      | Crash and reset after boot
| 6   | OK          | Fail      | Crash during game play
| 7   | OK          | Fail      | Crash and reset at high score screen
| 8   | OK          | Fail      | Freeze at ATLUS logo
| 9   | OK          | Fail      | Graphical glitch and crash, reset at title screen
| 10  | OK          | Fail      | Doesn't boot
| 11  | OK          | Fail      | Graphical glitch and crash, reset at title screen
| 12  | OK          | Fail      | Crash during game play
| 13  | OK          | Fail      | Crash during game play
| 14  | OK          | Fail      | Graphical glitch and crash, reset at title screen
| 15  | OK          | OK        | -
| 16  | OK          | Fail      | Crash during game play
| 17  | OK          | Fail      | Graphical glitch and crash, reset at title screen
| 18  | OK          | Fail      | Crash and reset after boot
| 19  | OK          | Fail      | Crash at Cave logo
| 20  | OK          | Fail      | Graphical glitch and freeze black screen after Cave logo


Only one build could run stably on Slow Case.
 
This indicates that SDRAM timing is not constrained well for 96 MHz.
And it can be said that it only works on some builds which happen to meet timing.


## Experiment 2: Fixed constraints

### Procedure
- Applied constraints to clocks as follows instead of decoupling all.

```
set_false_path -from [get_clocks $sys_clk] -to [get_clocks $video_clk]
set_false_path -from [get_clocks $video_clk] -to [get_clocks $sys_clk]

set_false_path -from [get_clocks $sdram_clk] -to [get_clocks $video_clk]
set_false_path -from [get_clocks $video_clk] -to [get_clocks $sdram_clk]

set_false_path -from [get_clocks $sdram_clk] -to [get_clocks $cpu_clk]
set_false_path -from [get_clocks $cpu_clk] -to [get_clocks $sdram_clk]

set_false_path -from [get_clocks $video_clk] -to [get_clocks $cpu_clk]
set_false_path -from [get_clocks $cpu_clk] -to [get_clocks $video_clk]
```

### Result

| No  | Normal Case | Slow Case | Failure detail
| --- | ----------- | --------- | --------------
| 1   | OK          | Fail      | Crash after boot

Applying constraints to clocks is not enough.
Still fails on Slow Case and got negative slack on Timing Analyzer.


## Experiment 3: Theoretical calculation of phase shift

### Procedure
- Calculated phase shift referring to "Embedded Peripherals IP User Guide" by Intel(Altera)

### Parameter
- SDRAM parameter
From datasheet of AS4C32M16SB-7TCN

| Parameter  | Value       | Description
| ---------- | ----------- | -----------------------
| `tOH`        | 2.5 ns      | Data output hold time
| `tDS`        | 1.5 ns      | Data/Address/Control Input set-up time
| `tDH`        | 0.8 ns      | Data/Address/Control Input hold time
| `tHZ`        | 5.4 ns      | Data output high impedance

- FPGA parameter
From "Datasheet Report" in Timing Analyzer

| Parameter  | Value       | Description
| ---------- | ----------- | -----------------------
| `tCLK`       | 10.417 ns   | Clock period (96 MHz)
| `tCO_MIN`    | 6.056 ns    | Minimum clock-to-output time  
| `tCO_MAX`    | 13.017 ns   | Maximum clock-to-output time
| `tH_MAX`    | -3.550 ns   | Maximum hold time after clock
| `tSU_MAX`    | 1.558 ns    | Maximum setup time before clock

### Result

- Read Lag
`Read Lag = tOH(SDRAM) - tH_MAX(FPGA) = 2.5 ns - (-3.550 ns) = 6.05 ns`

- Write Lag
`Write Lag = tCLK - tCO_MAX(FPGA) - tDS(SDRAM) = 10.417 ns - 13.017 ns - 1.5 ns = -4.1 ns`

- Read Lead
`Read Lead = tCLK - tHZ(SDRAM) - tSU_MAX(FPGA) = 10.417 ns  - 5.4 ns - 1.447 ns = 3.57 ns`

- Write Lead
`Write Lead = tCO_MIN(FPGA) - tDH(SDRAM) = 6.056 ns - 0.8 ns = 5.256 ns`

From the results of the calculation, phase shift should be satisfy the following relationship.
`4.1 ns < Phase Shift` and ` Phase Shift < 3.57 ns`

It's impossible to satisfy.
But this calculation uses worst case, so if it is relaxed, there should be pragmatic point around it.
About Write Lag, the smaller the `tCO_MAX` is, the more room there is in the lower limit.
And let's not forget about this calculation is just an estimation. 
It does not include delay on PCB and delay from the PLL clock output nodes to destinations.

FYI, current phase shift is 270 degrees (7813 ps).

## Experiment 4: Adjusting phase shift

### Procedure
- Did timing analysis using several phase shift value.
- To minimize delay on `SDRAM_A`, `SDRAM_BA` and `SDRAM_CKE`, added location assignment tweaks(see "Note and suggestion" section)
- Changed multicycle path constraints since phase shift is now less than 180 degrees.

### Result

| Shift     | Setup slack<br>(Slow 85, `clk_sys` to `clk_sdram`)  | Setup slack<br>(Fast 0, `clk_sdram` to `clk_sys`)
| --------- | :----------------------------------------------- | :--------------------------------------------
| +4166 ps  | negative on `SDRAM_A` (-0.791 to -0.28)           | positive
| +4296 ps  | negative on `SDRAM_A` (-0.6 to -0.2)              | negative on `SDRAM_DQ` (-0.1 to -0.02)
| +4427 ps  | negative on `SDRAM_A` (-0.26 to -0.129)           | negative on `SDRAM_DQ` (-0.26 to -0.129)
| +4817 ps  | negative on `SDRAM_A` (-0.5 to 0.09)              | negative on `SDRAM_DQ` (-0.492 - 0.398)

Hold slack was all positive on every corner.
Increasing the shift amount improves Setup slack(Slow 85, `clk_sys` to `clk_sdram`), but worsens the Setup slack(Fast 0, `clk_sdram to clk_sys`).
But as I say in "Note and suggestion" section, improving SDRAM Controller and assigning `SDRAM_A`, `SDRAM_BA` and `SDRAM_CKE` to fast output register will reduce Setup slack(Slow 85, `clk_sys` to `clk_sdram`) and widen the window.

## Experiment 5: Testing builds

### Procedure

- Selected +4166 ps as phase shift 
- Made 20 builds with diffrent fitter seed value using DSE II.

 There is no specific reason for selecting this value and it may not be perfect. 
 Only I can say is that it seems to be better than the current phase shift as a result in my environment. 
 I cannot decide what phase shift is perfect.

### Result

| No  | Normal Case | Slow Case
| --- | ----------- | ---------
| 1   | OK          | OK
| 2   | OK          | OK 
| 3   | OK          | OK
| 4   | OK          | OK
| 5   | OK          | OK
| 6   | OK          | OK
| 7   | OK          | OK
| 8   | OK          | OK
| 9   | OK          | OK
| 10  | OK          | OK
| 11  | OK          | OK
| 12  | OK          | OK
| 13  | OK          | OK
| 14  | OK          | OK
| 15  | OK          | OK
| 16  | OK          | OK
| 17  | OK          | OK
| 18  | OK          | OK
| 19  | OK          | OK
| 20  | OK          | OK

No failures.


## Experiment 6: (Extra) Using DDIO instead of phase shift

### Procedure

- Cherry picked a8666be on `sdram-ddio` branch into the latest `master` branch (and fixed conflicts).
- Made 20 builds with different fitter seed value using DSE II.

### Result

| No  | Normal Case | Slow Case | Failure detail
| --- | ----------- | --------- | --------------
| 1   | OK          | OK(?)     |
| 2   | OK          | OK(?)     |
| 3   | Failed      | Failed    | Corrupted color and sound.
| 4   | OK          | OK(?)     |
| 5   | Failed      | Failed    | Doesn't boot
| 6   | Failed      | Failed    | Doesn't boot
| 7   | Failed      | Failed    | Doesn't boot
| 8   | OK          | OK(?)     |
| 9   | OK          | OK(?)     |
| 10  | OK          | OK(?)     |
| 11  | OK          | OK(?)     |
| 12  | OK          | OK(?)     |
| 13  | OK          | OK(?)     |
| 14  | OK          | OK(?)     |
| 15  | OK          | OK(?)     |
| 16  | OK          | OK(?)     |
| 17  | OK          | OK(?)     |
| 18  | OK          | OK(?)     |
| 19  | OK          | OK(?)     |
| 20  | OK          | OK(?)     |


Both Normal Case and Slow Case had same issue on the sane build.
It indicates that the cause of the problem is not on forced delayed `SDRAM_DQ[1]` line.

And game was playable but explosion sound was cracked on all Slow Case.