@gnadenlos 2021-04-16
If you could upload the compiled core somewhere I could test it on my MiSTer. I am one of the users that had problems with the latest official release.