@bellwood420 2021-04-17
Thank you for your offer :)
It could be better to be tested in more environments since I cannot rule out the possibility of happening something bad in other environments for this kind of timing issue.

[_@CaveTest4166ps.zip](https://github.com/MiSTer-devel/Arcade-Cave_MiSTer/files/6328391/_%40CaveTest4166ps.zip)

Please put _@CaveTest4166ps directory into your SD root.
It contains 20 rbf files and 20 mra files for loading them.
Rom is excluded so put yours in mame directory.