@gnadenlos 2021-04-17
All of the 20 test files worked for me through level 1 (boss + score table). Couldn't test sound, because of the environment, but graphics were ok and I had no crashes.