@itsTrifle 2021-04-17
I also had issues with the 20210406 release, so downloaded the 4166ps test builds to try (128MB SDRAM, HDMI)
Have played through DDP level 1 boss to the start of level 2 on each of the rbfs and had no noticable sound or graphics issues.
Also checked Dangun Feveron and that is also working OK.
Thanks for your efforts!