@nullobject 2021-04-18
@bellwood420 Thanks so much for this really detailed analysis :heart: 

I actually can't build the project locally, as I keep getting fitter errors with this branch. Maybe it has something to do with the location assignments you added?

I suppose rather than try to fix it, I might just refactor the SDRAM controller to use registers for address, bank, etc. That way we don't even need to use the location assignments workaround.

I will merge this branch as soon as I make those changes.