@bellwood420 2021-04-18
@nullobject 

My bad. I also get the fitter error today.
When I change auto-generated `BUILD_DATE` macro string from `210418` to `210415`, it works.

Name and number of registers and LUTs generated by fitter in the end varies across builds with different pseudo-random seed including `BUILD_DATE` macro string.
So I used wildcard for location assignments, however, apparently the number of registers or LUTs generated today will not fit into the assigned cell.

It can be fixed by separating assignments into 2 different neighbor cells, but it is definitely better to fix the SDRAM controller.