@bellwood420 2021-04-19
Thanks for merging and updating SDRAM controller.

However I got the following warnings and address/bank registers are not assigned to fast I/O registers🤔

https://www.intel.com/content/www/us/en/programmable/quartushelp/13.0/mergedProjects/msgs/msgs/wfsac_fsac_register_packing_clr_load_violation.htm

![log](https://user-images.githubusercontent.com/50853463/115224597-05165900-a148-11eb-972d-ac5b3634fd95.png)