@nullobject 2021-04-19
@bellwood420 That sucks, there must be more to it than just using a register for bank and address :thinking: