@bellwood420 2021-04-19
@nullobject Awesome! Now they are assigned and clock-to-output time has improved.

![improvements](https://user-images.githubusercontent.com/50853463/115240621-e2da0680-a15a-11eb-84d6-2e517dff6ac4.png)